threadStarter and username comparison works now

+++ backend/app/controllers/thread.controller.js
@@ -1,31 +1,33 @@
 const config = require("../config/auth.config");
+const {
+  user
+} = require("../models");
 const db = require("../models");
 const Thread = db.thread;
 
 exports.postThread = (req, res) => {
-    // Validate request
-    if(!req.body) {
-        res.status(400).send({
-            message: "Content can not be empty"
-        });
-    }
-
-    // Create Thread
-    const thread = new Thread({
-        threadStarter: req.body.threadStarter,
-        category: req.body.category,
-        title: req.body.title
+  // Validate request
+  if (!req.body) {
+    res.status(400).send({
+      message: "Content can not be empty"
     });
+  }
 
-    // Save Thread in the database
-    Thread.create(thread, (err, data) => {
-        if(err)
-        res.status(500).send({
-            message:
-            err.message || "Some error occured while creating the Thread."
-        });
-        else res.send(data);
-    });
+  // Create Thread
+  const thread = new Thread({
+    threadStarter: req.body.threadStarter,
+    category: req.body.category,
+    title: req.body.title
+  });
+
+  // Save Thread in the database
+  Thread.create(thread, (err, data) => {
+    if (err)
+      res.status(500).send({
+        message: err.message || "Some error occured while creating the Thread."
+      });
+    else res.send(data);
+  });
 };
 
 exports.findAll = (req, res) => {
@@ -33,15 +35,16 @@ exports.findAll = (req, res) => {
   Thread.find(title, (err, data) => {
     if (err)
       res.status(500).send({
-        message:
-          err.message || "Some error occurred while retrieving Threads."
+        message: err.message || "Some error occurred while retrieving Threads."
       });
     else res.send(data);
   });
 };
 
 exports.findAllByCategory = (req, res) => {
-  Thread.find({category: req.params.category}, (err, data) => {
+  Thread.find({
+    category: req.params.category
+  }, (err, data) => {
     if (err) {
       if (err.kind === "not_found") {
         res.status(404).send({
@@ -56,7 +59,9 @@ exports.findAllByCategory = (req, res) => {
   });
 };
 exports.findUsersThreads = (req, res) => {
-  Thread.find({threadStarter: req.params.id})
+  Thread.find({
+      threadStarter: req.params.id
+    })
     .then(data => {
       res.send(data);
     })
@@ -68,17 +73,64 @@ exports.findUsersThreads = (req, res) => {
 }
 
 exports.delete = (req, res) => {
-    Thread.findByIdAndRemove(req.params.id, (err) => {
-      if (err) {
-        if (err.kind === "not_found") {
-          res.status(404).send({
-            message: `Not found Thread with id ${req.params.id}.`
+
+  Thread.findOne({
+    threadStarter: req.body.threadStarter
+  }).exec((err, thread) => {
+    if (err) {
+      res.status(500).send({
+        message: err
+      });
+      return;
+    }
+
+    if (thread) {
+
+      if (req.body.threadStarter == "") {
+        return res.status(400).send({
+          message: "threadStarter cannot be empty"
+        });
+      } else if(!req.body.threadStarter) {
+        return res.status(404).send({
+          message: "threadStarter does not exist..."
+        });
+      }
+      
+      else {
+        Thread.findByIdAndRemove(req.params.id, (err) => {
+          if (err) {
+            if (err.kind === "not_found") {
+              return res.status(404).send({
+                message: `Not found Thread.`
+              });
+            } else {
+              return res.status(500).send({
+                message: "Could not delete Thread with id "
+              });
+            }
+          } else res.send({
+            message: `Thread was deleted successfully!`
           });
-        } else {
-          res.status(500).send({
-            message: "Could not delete Thread with id " + req.params.id
-          });
-        }
-      } else res.send({ message: `Thread was deleted successfully!` });
-    });
-  };
+        });
+
+      }
+    }
+
+  })
+
+};
+
+
+// Thread.findByIdAndRemove(req.params.id, (err) => {
+//   if (err) {
+//     if (err.kind === "not_found") {
+//       res.status(404).send({
+//         message: `Not found Thread with id ${req.params.id}.`
+//       });
+//     } else {
+//       res.status(500).send({
+//         message: "Could not delete Thread with id " + req.params.id
+//       });
+//     }
+//   } else res.send({ message: `Thread was deleted successfully!` });
+// });
